[FIX] Update lastactivity at init
When the timeout option is used, we need to track the lastactivity for
each connection. This must be done immediatly at the init. Otherwise we
may fall in a timeout check with a zeroed timeval struct and the
connection will timeout immediatly. Subsequents event for this
connections are then lost.

Signed-off-by: Benjamin Hesmans <benjamin.hesmans@uclouvain.be>