Check SYN SYNACK key in third ACK
In some traces, the key sent by client or the server in SYN or SYNACK
may be different in the retransmissions.

We compare key observed in SYN and SYNACK with the one in the third
ACK. If they don't match, we take the keys from the third ack and
recompute the initial sequence number and update the hashtable of the
hashed keys to retrieve further joins.

Signed-off-by: Benjamin Hesmans <benjamin.hesmans@uclouvain.be>