gcc: fix 'implicit declaration of function' warnings
Simply by declaring functions in headers, include missing headers and
change functions order.

Easier to detect real issues when all these warnings go away.

While at it, isMPTCP_fastclose() was called with an extra argument. Now
that it is declared in the header file, we need to make sure the call is
aligned with the signature.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>